pyproject.toml
[tool.poetry]
name = "koni"
version = "0.1.0"
description = ""
authors = ["steven. <sveshkini@berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.25.0"
requests = "^2.28.1"
twilio = "^7.16.0"
Flask = "^2.2.2"
python-dotenv = "^0.21.0"
redis = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
